refactored markerLayer into routePathLinkLayer
