+++ src/components/map/layers/routePathLinkLayer.scss
@@ -1,0 +1,4 @@
+.nodeMarker {
+    display: block;
+    opacity: 0.8;
+}
